change test file location, minor fixes based on review comments
